+++ keyboards/tapedeck/keymaps/default/keymap.c
@@ -19,14 +19,6 @@
 
 #define BASE 0 
 
-#ifdef POT_ENABLE
-  #include "analog.h"
-#endif
-
-int16_t pot_val = 0;
-int16_t prev_val = 0;
-int16_t val = 0;
-
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
     [BASE] = LAYOUT(
         KC_ESC,            KC_MNXT,           KC_MPLY,     KC_MNXT, KC_MPLY,
@@ -41,44 +33,6 @@ const uint16_t PROGMEM encoder_map[][NUM_ENCODERS][2] = {
 };
 #endif
 
-void matrix_init_user(void) {
-#ifdef POT_ENABLE
-    analogReference(ADC_REF_POWER);
-#endif
-}
 
-uint8_t divisor = 0;
-
-long map(long x, long in_min, long in_max, long out_min, long out_max)
-{
-  return (x - in_min) * (out_max - out_min) / (in_max - in_min) + out_min;
-}
-
-void slider(void) {
-    if (divisor++) { // only run the slider function 1/256 times it's called
-        return;
-    }
-
-    pot_val = analogReadPin(D4);
-	val = map(pot_val, 0, 1023, 1, 50);
-
-	if (( val > (prev_val + 1)) && val != prev_val){
-		int i;
-		for (i = prev_val; i < val; i++ )
-				{tap_code(KC_VOLU);}
-	}
-	if ((val  < (prev_val - 1)) && val != prev_val){
-		int i;
-		for (i = val; i < prev_val; i++ )
-				{tap_code(KC_VOLD);}
-	}
-	prev_val = val;
-}
-
-void matrix_scan_user(void) {
-#ifdef POT_ENABLE
-	slider();
-#endif
-}
